Add files via upload
With these two files, I achieved my minimum viable product, software-wise. I set out to have a program that manipulated a sine tones frequency on the x-axis and its amplitude on its y-axis. I made the synth in a way that allows me to turn on and off this sine in times when we only want the drumbeat. The button plays a snare drum wave file. The Arduino file sends serial messages to supercollider which interprets them accordingly.